Move code from left/top sidebars to shared model
